Added KG-UVR5 radio and refactored KG-UV920A
